cluster: move offsets_snapshot.h from cloud_metadata to the upper level
It is going to be used for data migrations. While data migrations
themselves require cloud storage, offsets are being migrated via API
completely bypassing cloud storage.
