+++ src/v/cluster/BUILD
@@ -452,7 +452,6 @@ redpanda_cc_library(
         "cloud_metadata/key_utils.h",
         "cloud_metadata/manifest_downloads.h",
         "cloud_metadata/offsets_recovery_rpc_types.h",
-        "cloud_metadata/offsets_snapshot.h",
         "cloud_metadata/offsets_upload_rpc_types.h",
         "cloud_metadata/producer_id_recovery_manager.h",
         "cloud_metadata/types.h",
@@ -543,6 +542,7 @@ redpanda_cc_library(
         "notification_latch.h",
         "ntp_callbacks.h",
         "offsets_recovery_rpc_service.h",
+        "offsets_snapshot.h",
         "partition.h",
         "partition_balancer_backend.h",
         "partition_balancer_planner.h",
