+++ src/v/cluster/cloud_metadata/tests/offsets_recovery_test.cc
@@ -21,7 +21,6 @@
 #include "cluster/cloud_metadata/offsets_recoverer.h"
 #include "cluster/cloud_metadata/offsets_recovery_manager.h"
 #include "cluster/cloud_metadata/offsets_recovery_router.h"
-#include "cluster/cloud_metadata/offsets_snapshot.h"
 #include "cluster/cloud_metadata/offsets_upload_router.h"
 #include "cluster/cloud_metadata/offsets_uploader.h"
 #include "cluster/cloud_metadata/tests/cluster_metadata_utils.h"
@@ -29,6 +28,7 @@
 #include "cluster/cloud_metadata/uploader.h"
 #include "cluster/commands.h"
 #include "cluster/controller.h"
+#include "cluster/offsets_snapshot.h"
 #include "cluster/types.h"
 #include "config/node_config.h"
 #include "kafka/client/client.h"
